LayoutTests/fast/dom/insert-new-child-to-remove-during-mutation-events-crash-expected.txt

CONSOLE MESSAGE: HierarchyRequestError: The operation would yield an incorrect node tree.
This tests inserting a new node to be removed during mutation events of textContent setter. WebKit should not hit a crash.

PASS
